refactor: 트랜잭션 템플릿 분리
